refactor: use monaco provider classes
